Implement JWT into game state validation
The game now uses JSONWebTokens to store a token containing the user's
current game state on their device. This token is sent to the API when
starting the game, so as long as the user is playing the same scene,
they will be able to resume their existing session with characters
already found. Each character found updates the token and creates a new
one to store on the client's device. Each token has a unique id which
will be used to prevent multiple high-score submissions from the same
token.